Add timeout function to send() function
Per the requests library documentation, timeout can be added to prevent indefinite or long waits for server reply.  For Home Assistant, this would be helpful to prevent the log from filling up when a mythfrontend component is added but currently switched off - Home Assistant sends warnings if the media device doesn't respond within a given time (I believe 10 seconds).

Tested in command-line script, does not break backward compatibility with existing scripts.  Not adding the parameter seems to allow it to work as usual.